+++ backend/src/db/models/cost/index.js
@@ -1,0 +1,10 @@
+const mongoose = require("mongoose");
+
+const Schema = mongoose.Schema;
+
+const costSchema = new Schema({
+    text: String,
+    summa: Number
+  });
+
+  module.exports = Cost = mongoose.model('costs', costSchema);
